requirements.txt
torch
torchvision
numpy
pillow
scikit-image
matplotlib
tqdm